Add files via upload
The data was too big too import.  The dataset we worked was 1.1 million rows.  The dataset is complex with a change made to certain values of the survey that over a million consumers noted on a government website concerning financial organizations.  This is a helpful tool for companies not doing this type of analysis.  I hope to perfect the model and offer this as a Saas for mid sized financial organizations.